feat:Created exception class for services

+++ src/main/java/com/prodmanager/product/services/UserService.java
@@ -1,8 +1,13 @@
 package com.prodmanager.product.services;
 
 import com.prodmanager.product.entities.User;
+import com.prodmanager.product.exceptions.DatabaseException;
+import com.prodmanager.product.exceptions.ResourceNotFoundException;
 import com.prodmanager.product.repositories.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +25,7 @@ public class UserService {
 
     public User findUserById(Long id) {
         Optional<User> user = userRepository.findById(id);
-        return user.get();
+        return user.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insertUser(User user) {
@@ -28,13 +33,26 @@ public class UserService {
     }
 
     public User updateUser(Long id, User user) {
-        User entity = userRepository.getReferenceById(id);
-        updateData(entity, user);
-        return userRepository.save(entity);
+        try {
+            User entity = userRepository.getReferenceById(id);
+            updateData(entity, user);
+            return userRepository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("User cannot be deleted because it has associated orders or products.");
+        }
     }
 
     private void updateData(User entity, User user) {
